pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "smart-warehouse"
version = "0.1.0"
description = "Enterprise-ready multi-agent smart warehouse simulation platform"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
	{name = "Smart Warehouse Team"}
]

dependencies = [
	"paho-mqtt>=2.0.0",
	"pydantic>=2.6.0",
	"pydantic-settings>=2.2.0",
	"PyYAML>=6.0.0",
	"python-dotenv>=1.0.1",
	"fastapi>=0.110.0",
	"uvicorn[standard]>=0.27.0",
	"sqlalchemy[asyncio]>=2.0.27",
	"asyncpg>=0.29.0",
	"structlog>=24.1.0",
	"prometheus-client>=0.20.0",
	"opentelemetry-api>=1.24.0",
	"opentelemetry-sdk>=1.24.0",
	"opentelemetry-exporter-otlp>=1.24.0",
	"opentelemetry-instrumentation-fastapi>=0.45b0",
	"grpcio>=1.64.0",
	"grpcio-tools>=1.64.0",
	"googleapis-common-protos>=1.63.0",
	"aio-pika>=9.4.2"
]

[project.optional-dependencies]
dev = [
	"pytest>=8.1.0",
	"pytest-asyncio>=0.23.0",
	"ruff>=0.4.0",
	"mypy>=1.9.0",
	"pytest-mock>=3.12.0",
	"httpx>=0.27.0"
]

[tool.pytest.ini_options]
addopts = "-ra"
testpaths = ["tests"]

[tool.setuptools]
packages = ["smart_warehouse"]
